israel-news-module/processor/requirements.txt
openai
python-dotenv
python-dateutil
